# Optimize bfloat16 construction