### Ticket
[Link to Github Issue](https://github.com/tenstorrent/tt-metal/issues/15822)

### Problem description
bfloat16 construction speed is the current bottleneck during model loading in GGML.

### What's changed
Removes unnecessary runtime checks, casts and functions calls. As a result, time loading Tiny LLaMA 1.1B on GGML reduces from 15.2s to 13.4s

```plaintext
Before:
bin/llama-cli -ngl 30 -m tinyllama-1.1b-chat-v1.0.Q4_0.gguf -p "" -n 0 -nkvo   16.85s user 1.01s system 117% cpu 15.234 total

After:
bin/llama-cli -ngl 30 -m tinyllama-1.1b-chat-v1.0.Q4_0.gguf -p "" -n 0 -nkvo   15.17s user 0.86s system 119% cpu 13.437 total
```

### Checklist
- [ ] Post commit CI passes
- [ ] Blackhole Post commit (if applicable)
- [ ] Model regression CI testing passes (if applicable)
- [ ] Device performance regression CI testing passes (if applicable)
- [ ] **(For models and ops writers)** Full [new models](https://github.com/tenstorrent/tt-metal/actions/workflows/full-new-models-suite.yaml) tests passes
- [ ] New/Existing tests provide coverage for changes