@blozano-tt 2024-12-10
Rerunning CI: https://github.com/tenstorrent/tt-metal/actions/runs/12252764125