@blozano-tt 2024-12-10
> Also, I ended up implementing my own SIMD version to mass convert float into bfloat16. I am not sure how often TT needs to convert large amounts of FP32 into BFP16. But I imagine it's going to make some impact somewhere.

Thanks for sharing this.
I’ll keep it in mind in case I come across a matching use case in the code base.