@abhullar-tt 2024-12-11
> Also, I ended up implementing my own SIMD version to mass convert float into bfloat16. I am not sure how often TT needs to convert large amounts of FP32 into BFP16. But I imagine it's going to make some impact somewhere.
> 
> ```c
> static void internal_fp32_to_bf16(const float* x, bfloat16* y, size_t n) {
>     size_t i = 0;
> #if defined(__AVX512BF16__)
>       for (; i + 32 <= n; i += 32) {
>         _mm512_storeu_si512(
>             (__m512i *)(y + i),
>             __m512i(_mm512_cvtne2ps_pbh(_mm512_loadu_ps(x + i + 16),
>                                 _mm512_loadu_ps(x + i))));
>       }
> 
> #elif defined(__AVX2__)
>     // Process 8 floats at a time using AVX2
>     for (; i + 8 <= n; i += 8) {
>         __m256 fx = _mm256_loadu_ps(x + i);
>         __m256i ix = _mm256_castps_si256(fx);
>         
>         // Shift right by 16 bits to get the upper 16 bits of the float
>         ix = _mm256_srli_epi32(ix, 16);
>         
>         // Pack the 32-bit integers into 16-bit integers
>         __m128i iy = _mm256_cvtepi32_epi16(ix);
>         
>         // Store the result
>         _mm_storeu_si128((__m128i*)(y + i), iy);
>     }
> #elif defined(__SSE__)
>     for (i = 0; i + 4 <= n; i += 4) {
>         __m128 fx = _mm_loadu_ps(x + i);
>         __m128i ix = _mm_castps_si128(fx);
>         
>         // Shift right by 16 bits to get the upper 16 bits of the float
>         ix = _mm_srli_epi32(ix, 16);
>         
>         // Pack the 32-bit integers into 16-bit integers
>         ix = _mm_packus_epi32(ix, ix);
>         
>         // Store the result
>         _mm_storel_epi64((__m128i*)(y + i), ix);
>     }
> #endif
>     
>     // Handle remaining elements
>     for (; i < n; i++) {
>         uint32_t ix = *(const uint32_t*)(x + i);
>         y[i] = bfloat16(ix >> 16);
>     }
> }
> ```
> 
> Model load time is as follows (along with other optimizations).
> 
> ```
> scalar: 10.55s
> AVX512: 5.165s
> AVX2: 5.925s
> SSE: 6.91s
> ```

@TT-BrianLiu or @ayerofieiev-tt could this be helpful in ttnn tensor land?